refactor: change naming convention
