+++ src/components/membership-components/MembershipBillForm.tsx
@@ -4,7 +4,7 @@ import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select'
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
-import { AddPaymentMethodDialog } from '@/components/membership-components/AddPaymentMethodDialog'
+import { AddPaymentMethodDialog } from '@/components/membership-components/MembershipAddPaymentMethodDrawer'
 import { useQuery } from '@tanstack/react-query'
 import { getAuthenticatedApi } from '@/lib/api-client'
 import { paymentQueryKeys } from '@/lib/QueryKeys'
