+++ src/components/membership-components/MembershipDetailsDialog.tsx
@@ -26,10 +26,10 @@ import { SubscriptionApi } from '@/api/generated/apis/SubscriptionApi'
 import type { SubscriptionAvailedTableDTO } from '@/api/generated/models/SubscriptionAvailedTableDTO'
 import { getAuthenticatedApi } from '@/lib/api-client'
 import { useAuthSession } from '@/lib/auth/auth-session'
-import { useAddMemberDialogData } from '@/hooks/membership/useAddMemberData'
-import { AddBillingDialog } from './AddBillingForm'
-import { AddSubscriptionDialog } from './AddSubscriptionDialog'
-import { useMemberDetailsDialogData } from '@/hooks/membership/useMemberDetailsData'
+import { useAddMemberDialogData } from '@/hooks/membership/useMembershipDetails'
+import { AddBillingDialog } from './MembershipBillForm'
+import { AddSubscriptionDialog } from './MembershipAddSubscriptionDrawer'
+import { useMemberDetailsDialogData } from '@/hooks/membership/useMembershipCreationContext'
 import { isAdminSession } from '@/lib/auth/auth-permissions'
 import { useBillingActions } from '@/hooks/billing/useBillingActions'
 import { memberQueryKeys } from '@/lib/QueryKeys'
