+++ src/hooks/membership/useAddMembership.ts
@@ -8,7 +8,7 @@ import { MemberApi } from '@/api/generated/apis/MemberApi'
 import { SubscriptionApi } from '@/api/generated/apis/SubscriptionApi'
 import type { SubscriptionAvailedTableDTO } from '@/api/generated/models/SubscriptionAvailedTableDTO'
 import { getAuthenticatedApi } from '@/lib/api-client'
-import { useAddMemberDialogData } from '@/hooks/membership/useAddMemberData'
+import { useAddMemberDialogData } from '@/hooks/membership/useMembershipDetails'
 import { useBillingActions } from '@/hooks/billing/useBillingActions'
 import { memberQueryKeys } from '@/lib/QueryKeys'
 import { apiResponseMemberTableSchema, memberPostDtoSchema } from '@/types/membership/memberSchemas'
