+++ src/hooks/membership/useMembership.ts
@@ -1,7 +1,7 @@
 import { useCallback, useMemo } from 'react'
 import { useQuery } from '@tanstack/react-query'
 import { useInvoices } from '@/hooks/billing/useInvoices'
-import { useMemberSubscriptions } from './useMemberSubscriptions'
+import { useMemberSubscriptions } from './useMembershipSubscriptions'
 import { useSubscriptionAvailed } from './useSubscriptionAvailed'
 import { readPersistedAuthToken } from '@/lib/auth/auth-session'
 import { apiResponseListMemberTableSchema } from '@/types/membership/memberSchemas'
