+++ src/hooks/membership/useMembershipCreationContext.ts
@@ -1,0 +1,11 @@
+import { useSubscriptionAvailed } from './useSubscriptionAvailed'
+import { useMemberSubscriptionByActorId } from './useMembershipSubscriptions'
+
+export function useMemberDetailsDialogData(options: { open: boolean; memberActorId: string | null }) {
+  const { open, memberActorId } = options
+
+  const subscriptionsQuery = useSubscriptionAvailed(open)
+  const memberSubscriptionQuery = useMemberSubscriptionByActorId(memberActorId, open)
+
+  return { subscriptionsQuery, memberSubscriptionQuery }
+}
