+++ src/routes/dashboard/marketing/membership.tsx
@@ -5,9 +5,9 @@ import { format } from 'date-fns'
 import { Search, QrCode, Fingerprint, UserCheck, Clock } from 'lucide-react'
 
 import type { MemberFormData, MemberInfo, AttendanceRecord, MembershipSearchForm } from '@/types/membership/memberSchemas'
-import MembersTable from '@/components/membership-components/MembersTable'
-import { MemberDetailsDialog } from '@/components/membership-components/MemberDetailsDialog'
-import { useMembersData } from '@/hooks/membership/useMembers'
+import MembersTable from '@/components/membership-components/MembershipTable'
+import { MemberDetailsDialog } from '@/components/membership-components/MembershipDetailsDialog'
+import { useMembersData } from '@/hooks/membership/useMembership'
 import { Input } from '@/components/ui/input'
 import { Button } from '@/components/ui/button'
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '@/components/ui/table'
